test: fix failing tests, and break apart tests into pages
