+++ integration-tests/gatsby-image/cypress/integration/fluid.js
@@ -2,7 +2,7 @@ const fluidTestId = `image-fluid`
 
 describe(`fluid`, () => {
   beforeEach(() => {
-    cy.visit(`/`).waitForRouteChange()
+    cy.visit(`/fluid`).waitForRouteChange()
   })
 
   it(`renders a spacer div`, () => {
