+++ integration-tests/gatsby-image/cypress/integration/traced.js
@@ -2,14 +2,19 @@ const tracedTestId = `image-traced`
 
 describe(`fixed`, () => {
   beforeEach(() => {
-    cy.visit(`/`).waitForRouteChange()
+    cy.visit(`/traced`).waitForRouteChange()
   })
 
-  // TODO: test this further
   it(`renders a traced svg`, () => {
     cy.getTestElement(tracedTestId)
       .find(`.gatsby-image-wrapper > img`)
       .should(`have.attr`, `src`)
-      .and(`contains`, `data:image/svg+xml`)
+      .and(src => {
+        [
+          `data:image/svg+xml`,
+          `fill='white'`
+        ]
+          .forEach(part => expect(src).to.include(part) )
+      })
   })
 })
